+++ kitty/config.py
@@ -94,7 +94,9 @@ def goto_tab_parse(func, rest):
 
 @func_with_args('detach_window')
 def detach_window_parse(func, rest):
-    return func, to_cmdline(rest)
+    if rest not in ('new', 'new-tab', 'ask'):
+        rest = 'new'
+    return func, (rest,)
 
 
 @func_with_args('set_background_opacity', 'goto_layout', 'kitty_shell')